plag search now has the option to read candret from file

+++ src/main/java/no/roek/nlpgraphs/document/PlagiarismPassage.java
@@ -8,11 +8,11 @@ public class PlagiarismPassage {
 
 	private String trainFile, testFile;
 	private int trainSentence, testSentence;
-	private float similarity;
+	private double similarity;
 	private Graph testGraph, trainGraph;
 	private String dir, testDir, trainDir;
 	
-	public PlagiarismPassage(ConfigService cs, String trainFile, int trainSentence, String testFile, int testSentence, float similarity) {
+	public PlagiarismPassage(ConfigService cs, String trainFile, int trainSentence, String testFile, int testSentence, double similarity) {
 		this.dir = cs.getParsedFilesDir();
 		this.testDir = cs.getTestDir();
 		this.trainDir = cs.getTrainDir();
@@ -70,7 +70,7 @@ public class PlagiarismPassage {
 		this.testSentence = testSentence;
 	}
 
-	public float getSimilarity() {
+	public double getSimilarity() {
 		return similarity;
 	}
